backend/requirements.txt
aiosmtplib==2.0.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
bcrypt==4.1.2
blinker==1.8.1
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
distlib==0.3.8
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email-validator==2.1.0
fastapi==0.110.0
fastapi-mail==1.4.1
filelock==3.14.0
Flask==3.0.3
fonttools==4.59.0
greenlet==3.2.3
h11==0.16.0
httptools==0.6.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.2
kiwisolver==1.4.8
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow-sqlalchemy==0.29.0
matplotlib==3.10.5
mysql-connector==2.2.9
numpy==2.3.2
packaging==25.0
pandas==2.3.1
passlib==1.7.4
pillow==11.3.0
pipenv==2023.12.1
platformdirs==4.2.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
PyJWT==2.8.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rsa==4.9.1
setuptools==69.5.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlparse==0.5.0
starlette==0.36.3
ta==0.11.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.1
urllib3==2.5.0
uvicorn==0.29.0
virtualenv==20.26.1
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.0.2
WTForms==3.1.2
